Add sign-in and sign-out buttons
